Use netty threads for handling scala future callbacks in Hub
